Added pagination enum instead of args in functions to better handle ranges of data.

+++ Sources/DCoreKit/Net/Model/Request/GetAccountHistory.swift
@@ -7,7 +7,7 @@ struct GetAccountHistory<Input>: BaseRequestConvertible where Input: Operation {
     
     init(_ accountId: ChainObject,
          stopId: ChainObject = ObjectType.operationHistoryObject.genericId,
-         limit: Int = 100,
+         limit: UInt64 = 100,
          startId: ChainObject = ObjectType.operationHistoryObject.genericId
         ) {
