+++ Sources/DCoreKit/Net/Model/Request/GetRelativeAccountHistory.swift
@@ -5,7 +5,7 @@ struct GetRelativeAccountHistory<Input>: BaseRequestConvertible where Input: Ope
     typealias Output = [OperationHistory<Input>]
     private(set) var base: BaseRequest<[OperationHistory<Input>]>
     
-    init(_ accountId: ChainObject, stop: Int = 0, limit: Int = 100, start: Int = 0) {
+    init(_ accountId: ChainObject, stop: UInt64 = 0, limit: UInt64 = 100, start: UInt64 = 0) {
         
         precondition(accountId.objectType == .accountObject, "Not a valid account object id")
         self.base = GetRelativeAccountHistory.toBase(
